rename hideProcess to hide_process and make changes to source.py
